Added @2x versions of images. Upgraded TouchXML. Fixed positioning to work with 3.5" and 4" displays. Fixed code warnings.

+++ Tedometer/Source/CostMeter.m
@@ -126,7 +126,7 @@ static NSNumberFormatter *tickLabelStringNumberFormatter;
 	if( mtuNumber == 0 ) 
 		parentNodePath = @"Cost.Total";
 	else 
-		parentNodePath = [NSString stringWithFormat: @"Cost.MTU%d", mtuNumber];
+		parentNodePath = [NSString stringWithFormat: @"Cost.MTU%ld", (long)mtuNumber];
 	
 	isSuccessful = [TedometerData loadIntegerValuesFromXmlDocument:document intoObject:self withParentNodePath:parentNodePath 
 										   andNodesKeyedByProperty:nodesKeyedByProperty];
